requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
async-timeout==5.0.1
billiard==4.2.3
celery==5.3.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2025.6
cryptography==46.0.3
cssbeautifier==1.15.4
cssselect2==0.8.0
Django==4.2.7
django-appconf==1.2.0
django-cors-headers==4.3.1
django-crispy-forms==2.5
django-debug-toolbar==4.2.0
django-encrypted-model-fields==0.6.5
django-environ==0.12.0
django-filter==25.1
django-formtools==2.5.1
django-mfa3==1.1.0
django-ratelimit==3.0.1
django-tenants==3.9.0
django_compressor==4.6.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
djlint==1.36.4
drf-yasg==1.21.11
EditorConfig==0.17.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
factory-boy==3.3.0
Faker==38.2.0
fido2==2.0.0
freetype-py==2.5.1
gunicorn==21.2.0
html5lib==1.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsbeautifier==1.15.4
json5==0.12.1
kombu==5.6.1
lxml==6.0.2
openpyxl==3.1.2
oscrypto==1.3.0
packaging==25.0
pathspec==0.12.1
pdfkit==1.0.0
Pillow==10.0.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.7
pycairo==1.29.0
pycparser==2.23
Pygments==2.19.2
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pyotp==2.9.0
pypdf==6.4.2
pytest==9.0.1
pytest-django==4.5.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
qrcode==7.3.1
razorpay==2.0.0
rcssmin==1.2.2
redis==5.0.1
regex==2025.11.3
reportlab==4.4.6
requests==2.32.5
rjsmin==1.2.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.4
svglib==1.6.0
tinycss2==1.5.1
tomli==2.3.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.17
